Integrate LazyFrames with memory-intensive agents.
- LazyFrames is now used by DQN and ACER.
- Add `img_inputs` as BaseAgent flag to automatically determine input scaling.
- Merge both tf and numpy batch dtypes into a list of strings `batch_dtypes` and revert concat_buffer_samples() and its test to previous version.
- Update tests accordingly.


Former-commit-id: 40013ea176a3fab3553b65fb91d72cc77f7a3401

+++ xagents/td3/agent.py
@@ -83,7 +83,9 @@ class TD3(DDPG):
             new_actions = tf.clip_by_value(
                 self.target_actor(new_states) + noise, -1.0, 1.0
             )
-            target_critic_input = tf.concat([new_states, new_actions], 1)
+            target_critic_input = tf.concat(
+                [tf.cast(new_states, tf.float64), tf.cast(new_actions, tf.float64)], 1
+            )
             target_value1 = self.target_critic1(target_critic_input)
             target_value2 = self.target_critic2(target_critic_input)
             target_value = tf.minimum(target_value1, target_value2)
